# openAPI/public-api.yaml
...
openapi: 3.0.3
info:
  title: "SIS Public API"
  description: "Public API for the Stored Identity Service"
  version: "1.0.0"
paths:
...
  /confirm-details:
    get:
      parameters:
        - $ref: "#/components/parameters/RedirectUriParam"
        - $ref: "#/components/parameters/StateParam"
        - $ref: "#/components/parameters/AuthCodeParam"
      responses:
        "200":
          description: OK
          content:
            text/html:
              schema:
                type: string
      x-amazon-apigateway-integration:
        httpMethod: POST
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${GetConfirmDetailsHandler.Arn}:live/invocations
        credentials:
          Fn::GetAtt: GetConfirmDetailsHandlerRestApiRole.Arn
        type: aws_proxy
    post:
      parameters:
        - in: query
          name: redirect_uri
          schema:
            type: string
      responses:
        "302":
          description: OK
          content:
            text/html:
              schema:
                type: string
      x-amazon-apigateway-integration:
        httpMethod: POST
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${PostConfirmDetailsHandler.Arn}:live/invocations
        credentials:
          Fn::GetAtt: PostConfirmDetailsHandlerRestApiRole.Arn
        type: aws_proxy
...
components:
  parameters:
...
    RedirectUriParam:
      in: query
      name: redirect_uri
      required: true
      description: |
        This value must exactly match one of the redirect uris registered by the Relying Party service and must be
        URL-encoded.
      schema:
        type: string
        format: uri
...
    StateParam:
      in: query
      name: state
      required: true
      description: |
        This value is used to validate the response sent to the redirect URI.  This value will be
        returned to the client in the authentication response.
      schema:
        type: string
    AuthCodeParam:
      in: query
      name: code
      required: true
      description: |
        The auth code.
      schema:
        type: string
...